refactor: move anchor transformation logic into external class
