+++ editor/src/main/java/com/github/kitakkun/noteapp/editor/EditorModule.kt
@@ -1,6 +1,7 @@
 package com.github.kitakkun.noteapp.editor
 
 import androidx.navigation.NavController
+import com.github.kitakkun.noteapp.editor.usecase.AnchorTransformer
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.core.module.dsl.factoryOf
 import org.koin.dsl.module
@@ -13,7 +14,9 @@ val editorModule = module {
             navController = navController,
             historyManager = get(),
             settingDataStore = get(),
+            anchorTransformer = get(),
         )
     }
     factoryOf(::EditHistoryManager)
+    factoryOf(::AnchorTransformer)
 }
